Refactor documentation tasks with shared abstract base class
Extract common properties and methods into AbstractDocumentTask to reduce duplication and improve maintainability. Update WhcTask and AbstractTransformationTask to extend the new base class, streamlining logging configuration and shared behavior.
