+++ build-logic/src/main/groovy/org/omegat/documentation/AbstractDocumentTask.groovy
@@ -1,0 +1,40 @@
+package org.omegat.documentation
+
+import org.gradle.api.DefaultTask
+import org.gradle.api.file.DirectoryProperty
+import org.gradle.api.logging.LogLevel
+import org.gradle.api.tasks.InputDirectory
+import org.gradle.api.tasks.Internal
+
+abstract class AbstractDocumentTask  extends DefaultTask implements DocConfigurable {
+
+    @InputDirectory
+    final DirectoryProperty styleDir = project.objects.directoryProperty()
+
+    /**
+     * Sources root for the documentation
+     */
+    @Internal
+    final DirectoryProperty docRoot = project.objects.directoryProperty()
+
+    @Internal
+    final DirectoryProperty outputRoot = project.objects.directoryProperty()
+
+    @Override
+    void configureWith(DocConfigExtension extension) {
+        docRoot.convention(extension.docRoot)
+        styleDir.convention(extension.styleDir)
+        outputRoot.convention(extension.outputRoot)
+    }
+
+    protected void configureLogging() {
+        // Redirect spurious task output to INFO unless explicitly configured for debug
+        switch (project.gradle.startParameter.logLevel) {
+            case LogLevel.DEBUG:
+                break
+            default:
+                logging.captureStandardOutput(LogLevel.INFO)
+                logging.captureStandardError(LogLevel.INFO)
+        }
+    }
+}
