+++ build-logic/src/main/groovy/org/omegat/documentation/WhcTask.groovy
@@ -2,35 +2,21 @@ package org.omegat.documentation
 
 import com.xmlmind.util.StringUtil
 import groovy.transform.CompileStatic
-import org.gradle.api.DefaultTask
 import org.gradle.api.file.Directory
-import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.FileTree
 import org.gradle.api.file.RegularFile
-import org.gradle.api.logging.LogLevel
 import org.gradle.api.provider.ListProperty
 import org.gradle.api.provider.Provider
 import org.gradle.api.tasks.Input
-import org.gradle.api.tasks.InputDirectory
 import org.gradle.api.tasks.InputFile
 import org.gradle.api.tasks.InputFiles
-import org.gradle.api.tasks.Internal
 import org.gradle.api.tasks.OutputDirectory
 import org.gradle.api.tasks.TaskAction
 import com.xmlmind.whc.Compiler
 import org.gradle.api.tasks.options.Option
 
 @CompileStatic
-class WhcTask extends DefaultTask implements DocConfigurable {
-
-    @InputDirectory
-    final DirectoryProperty styleDir = project.objects.directoryProperty()
-
-    @Internal
-    final DirectoryProperty docRoot = project.objects.directoryProperty()
-
-    @Internal
-    final DirectoryProperty outputRoot = project.objects.directoryProperty()
+class WhcTask extends AbstractDocumentTask {
 
     @InputFile
     final Provider<RegularFile> inputFile = project.objects.fileProperty()
@@ -50,13 +36,8 @@ class WhcTask extends DefaultTask implements DocConfigurable {
 
     @TaskAction
     void transform() {
-        switch (project.gradle.startParameter.logLevel) {
-            case LogLevel.DEBUG:
-                break
-            default:
-                logging.captureStandardOutput(LogLevel.INFO)
-                logging.captureStandardError(LogLevel.INFO)
-        }
+        configureLogging()
+
         Compiler compiler = new Compiler(null)
         compiler.setVerbose(true)
         File[] contents = contentFiles.get().getFiles().toArray(new File[0])
@@ -66,10 +47,4 @@ class WhcTask extends DefaultTask implements DocConfigurable {
         compiler.compile(contents, tocFile.get().asFile, inputFile.get().asFile, outputDirectory.get().asFile)
     }
 
-    @Override
-    void configureWith(DocConfigExtension extension) {
-        docRoot.convention(extension.docRoot)
-        outputRoot.convention(extension.outputRoot)
-        styleDir.convention(extension.styleDir)
-    }
 }
